+++ src/server/trpc/context.ts
@@ -19,30 +19,35 @@ export const createContextInner = async (opts: CreateInnerContextOptions) => {
 };
 
 export const createContext = async (opts: createSolidAPIHandlerContext) => {
-	// Only required in non-GET requests (POST, PUT, DELETE, PATCH, etc)
-	const originHeader = opts.req.headers.get("Origin");
-	// NOTE: You may need to use `X-Forwarded-Host` instead
-	const hostHeader = opts.req.headers.get("Host");
-	if (
-		!originHeader ||
-		!hostHeader ||
-		!verifyRequestOrigin(originHeader, [hostHeader])
-	) {
-		//BUG see if you can even return this response in the first place
-		return new Response(null, {
-			status: 403,
-		});
+	if (import.meta.env.PROD) {
+		// Only required in non-GET requests (POST, PUT, DELETE, PATCH, etc)
+		const originHeader = opts.req.headers.get("Origin");
+		// NOTE: You may need to use `X-Forwarded-Host` instead
+		const hostHeader = opts.req.headers.get("Host");
+		if (
+			!originHeader ||
+			!hostHeader ||
+			!verifyRequestOrigin(originHeader, [hostHeader])
+		) {
+			//BUG see if you can even return this response in the first place
+			console.log("test1");
+			return new Response(null, {
+				status: 403,
+			});
+		}
 	}
 
 	const cookieHeader = opts.req.headers.get("Cookie");
 	const sessionId = lucia.readSessionCookie(cookieHeader ?? "");
 	if (!sessionId) {
+		console.log("test2");
 		return new Response(null, {
 			status: 401,
 		});
 	}
 	//BUG does this new headers thing work here
 	const headers = new Headers();
+	console.log(headers);
 
 	const { session, user } = await lucia.validateSession(sessionId);
 	if (!session) {
